Refactor CoverEditor component
Refactor CoverEditor component into smaller, more manageable files.
